Don’t scale XYZ by 100 to avoid needless floating point operations
The XYZ coordinates of the colour are only ever used internally
by the library and never exposed to the customer.  Still, the
coordinates are scaled by a hundred only later to be divided (in
case of Y by a hundred).

Do not scale the coordinates to avoid performing needless floating
point operations.  In case of Y coordinate, multiplication and
division is removed.  In case of X and Z, multiplication is removed
and division is scalled appropriatly to give the same result.

Reducing number of time floating points are manipulated increases
accuracy which can be observed looking at rgb(196, 196, 196) which
is now correctly converted to La*b* colour space with a* and b*
components equal zero.