Fix E₀ threshold when doing gamma expansion
The sRGB gamma expansion function has a linear and exponential parts.
The function switches from one to the other at argument equal to
around 0.0404.  However, the `rgb_to_xyz_map` function uses threshold
equal 10.  This is fine if the argument is an integer in 0–255
range (255 * 0.0404 = 10.302) which the function was originally
designed to handle.  However, it has since been changed to accept
floating point numbers in said range.  As a result, if a colour
component is, say, 10.1, the function will use the wrong gamma
expansion formula.

Since Lab::from_rgb_normalized is a hidden interface, chances are that
no one actually run into this issue.  The issue is also not present
when doing gamma compression (as is done when going from XYZ to sRGB).

Nonetheless, address the issue by using exact E₀ value (scaled
by 255).  While at it, update value for S₀ as well.